Move traits::query datatypes to traits::types.
